+++ nexus/lib/core/wandbcore.go
@@ -7,146 +7,22 @@ import (
 	_ "embed"
 	"errors"
 	"fmt"
-	"log"
 	"os"
-	"os/exec"
-	"runtime"
 	"strconv"
 	"strings"
-	"sync"
-	"syscall"
 	"time"
-	"unsafe"
 
 	"context"
 	"flag"
 
 	"github.com/wandb/wandb/nexus/pkg/client"
+	"github.com/wandb/wandb/nexus/internal/execbin"
 )
 
 // generate nexus binary and embed into this package
 //go:generate go build -C ../.. -o lib/core/libwandbcore.bin cmd/nexus/main.go
 //go:embed libwandbcore.bin
-
 var filePayload []byte
-
-func MemfdCreate(path string) (r1 uintptr, err error) {
-	s, err := syscall.BytePtrFromString(path)
-	if err != nil {
-		return 0, err
-	}
-
-	r1, _, errno := syscall.Syscall(319, uintptr(unsafe.Pointer(s)), 0, 0)
-
-	if int(r1) == -1 {
-		return r1, errno
-	}
-
-	return r1, nil
-}
-
-func CopyToMem(fd uintptr, buf []byte) (err error) {
-	_, err = syscall.Write(int(fd), buf)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func ExecveAt(fd uintptr) (err error) {
-	s, err := syscall.BytePtrFromString("")
-	if err != nil {
-		return err
-	}
-	// port-filename
-	argv := []string{"wandb-core", "--port-filename", "junk-pid.txt"}
-	envv := os.Environ()
-	// argv0p, err := BytePtrFromString(argv0)
-	// if err != nil {
-	// 	return err
-	// }
-	argvp, err := syscall.SlicePtrFromStrings(argv)
-	if err != nil {
-		return err
-	}
-	envvp, err := syscall.SlicePtrFromStrings(envv)
-	if err != nil {
-		return err
-	}
-	ret, _, errno := syscall.Syscall6(322, fd, uintptr(unsafe.Pointer(s)),
-		uintptr(unsafe.Pointer(&argvp[0])),
-		uintptr(unsafe.Pointer(&envvp[0])),
-		0x1000 /* AT_EMPTY_PATH */, 0)
-	if int(ret) == -1 {
-		return errno
-	}
-
-	// never hit
-	log.Println("should never hit")
-	return err
-}
-
-func run_syscall() {
-	fd, err := MemfdCreate("/file.bin")
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = CopyToMem(fd, filePayload)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = ExecveAt(fd)
-	if err != nil {
-		log.Fatal(err)
-	}
-}
-
-func run_file() *exec.Cmd {
-	file, err := os.CreateTemp("", "wandb-core-")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer os.Remove(file.Name())
-	_, err = file.Write(filePayload)
-	if err != nil {
-		log.Fatal(err)
-	}
-	file.Close()
-	err = os.Chmod(file.Name(), 0500)
-	if err != nil {
-		log.Fatal(err)
-	}
-	cmd := exec.Command(file.Name(), "--port-filename", "junk-pid.txt")
-	cmd.Env = os.Environ()
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err = cmd.Start()
-	if err != nil {
-		switch e := err.(type) {
-		case *exec.Error:
-			fmt.Println("failed executing:", err)
-		case *exec.ExitError:
-			fmt.Println("command exit rc =", e.ExitCode())
-		default:
-			panic(err)
-		}
-	}
-	fmt.Printf("write %+v\n", file.Name())
-	return cmd
-}
-
-func handleReader(reader *bufio.Reader) {
-	for {
-		str, err := reader.ReadString('\n')
-		if err != nil {
-			break
-		}
-		fmt.Print(str)
-	}
-}
 
 // readLines reads a whole file into memory
 // and returns a slice of its lines.
@@ -170,7 +46,7 @@ func tryport() (int, error) {
 	if err != nil {
 		return 0, err
 	}
-	if len(lines) != 2 {
+	if len(lines) < 2 {
 		return 0, errors.New("can't work with 1")
 	}
 	pair := strings.SplitN(lines[0], "=", 2)
@@ -233,68 +109,28 @@ func junk() {
 	}
 }
 
-func fork_exec() {
-	foo := 4
-	bar := 10
-	id, _, _ := syscall.Syscall(syscall.SYS_FORK, 0, 0, 0)
-	if id == 0 {
-		foo++
-		// fmt.Println("In child:", id, foo, bar)
-		run_syscall()
-	} else {
-		bar++
-		junk()
-		// TODO wait for child to exit
-		os.Exit(0)
-	}
-}
-
 //export wandbcore_setup
 func wandbcore_setup() {
 	// run()
 }
 
-func waitcmd(command *exec.Cmd) error {
-	if err := command.Wait(); err != nil {
-		if exiterr, ok := err.(*exec.ExitError); ok {
-			if status, ok := exiterr.Sys().(syscall.WaitStatus); ok {
-				fmt.Printf("Exit Status: %+v\n", status.ExitStatus())
-				return err
-			}
-		}
-		return err
-	}
-	return nil
-}
-
 //export wandbcore_init
 func wandbcore_init() int {
 	os.Remove("junk-pid.txt")
-	var cmd *exec.Cmd
-	// TODO: use build constraints (check for amd64 too prob)
-	if runtime.GOOS == "linux" {
-		fork_exec()
-	} else {
-		cmd = run_file()
-	}
-	junk()
-	if cmd != nil {
-		err := waitcmd(cmd)
-		if err != nil {
-			panic(err)
-		}
-	}
-	os.Exit(0)
 
-	wg := sync.WaitGroup{}
-	wg.Add(1)
-	go func() {
-		fmt.Printf("out\n")
-		time.Sleep(10 * time.Second)
-		fmt.Printf("done\n")
-		wg.Done()
-	}()
-	wg.Wait()
+	args := []string{"--port-filename", "junk-pid.txt"}
+	cmd, err := execbin.ForkExec(filePayload, args)
+	if err != nil {
+		panic(err)
+	}
+
+	// TODO: dont do this
+	junk()
+
+	err = cmd.Wait()
+	if err != nil {
+		panic(err)
+	}
 	return 22
 }
